use generic bitset-based set intersection instead of "smart" handling
Smart handling (as done before) cannot be used when parallel and single-node hyperedges
are eliminated during coarsening, because this invalidates the invariant necessary for
the previous uncontraction approach to work (namely that the "last couple of entries" that
are new to the current state are the case 2 ones). Since this can't be inferred from the
position anymore (because both parallel-edge and single-node-edge removal premute the pinlist
to account for removed edges), we have to use a generic approach and actually intersect both sets.

We might look into different approaches to do this. The current one should not be seen as final!