bugfix: always create new incidence information when HE is added to rep during contraction
Previous version tried to not waste memory in case the incidence information of
the representative hypernode is already at the end of the _incidence_array. If this
was the case, the old entries were not copied and the new ones where just appended at the
end. This behavior is not problematic if we do not care about the removal of parallel and
single-node HEs. However we do care and remove any parallel and single-node HEs. In the
concrete example of ibm03, this lead to incorrect uncontraction behavior:

Both types of hyperedges are removed from the hypernode by swapping their correspondiong
incidence entrie to the end of the incidence array of the hypernode and then decreasing
the size of the hypernode-vertex. Thus the incidence entries of u might get manipulated
after the contraction (i.e. in the course of single-node and parallel HE removal).

Now, assume that the incidence entries of a representative hypernode u are already at the
end of the incidence_array and the current contraction operation adds a few more hyperedges
to u. In the previous case, these entries were just appended to the end of the _incidence_array.
If no parallel or single-node HEs are removed after the contraction, everything is fine.
However if we DO remove HEs by permuting the incidences entries of u, we effectively destroy
the history information of the contraction memento, because the memento only stores pointers
to the first of the old entries and the size of the old entries before contraction.

Now because of the fact that both the "new" and the "old" incidence array of u have the same
beginning (i.e. start at the same index in _incidence_array) and the new one is just longer as
the old one (i.e. the old one is a subset of the new one), removal of parallel and single-node
HEs operates one both incidence arrays instead of only modifying the new one.

If an uncontraction operation now tries to retrieve the old entries, by using the first_entry
and size information of the contraction memento, it might look at other entries than the
original ones, because the original ones might have been swapped to another position during
the removal of parallel or single-node HEs. Thus in this case, it is not possible anymore to
infer which entries lead to a case 1 (aka update operation) or case 2 (aka append operation)
during contraction and thus the uncontraction operation fails.

This commit now fixes the problem by always ensuring that a contraction operation during which
new hyperedges are added to the representative lead to new entries in the _incidence_array -
even if the entries of the hypernode in question are already at the end of the _incidence_array.
Thus it now consistently maintains both old (before contraction) and new (after contraction)
state of the representative hypernode.
During uncontraction this now ensures that it is possible to distinguish between case 1 and
case 2 operations and therefore apply the correct uncontraction operations for the hyperedges
in question!