fix updating of neighbour-nodes by remembering just activated nodes
Previous version contained a bug in updateNeighbours which lead to
incorrect gain calculation. Applying delta-Gain changes is only correct for
nodes that were already contained in the PQ when the max_gain node was
moved (e.g. it is not correct for newly added (i.e. activated) nodes).
However the updatePin method also accounts for nodes that become
activated due to the current node movement: These nodes are activated
and inserted into the PQ. However, if such a node is incident to more
than one HE that is considered during updateNeighbours, this leads
to the implicit assumption that the node should be updated via
delta-gain updates (since it was only checked whether or not the node is
already in the PQ and not whether or not the node was already active
before the current max_gain_node movement).

Bug was found by assertion that the current cut (calculated by adding
the gain changes) has to match the cut induced by the current hypergraph
configuration.

This commit fixes this bug by introducing a bitvector to remember newly
added nodes which are then excluded from delta gain update.