change hypergraph behavior: unconnected hypernodes are removed
Previous version left hypernodes enabled after removing the last
incident hyperedge. During coarsening this lead to the problem that
the number of hypernodes for initial partitioning could not be reduced
to the predefined threshold since unconnected hypernodes returned an
<<invalid>> rating.

Since we currently do not benefit from having unconnected hypernodes
in the graph, we disable them as soon as the last incident hyperedge
is removed and re-enable them if the edge is restored.

Testcase in hypergraph_test has been adapted to reflect this behavior.
Testcases in the coarsener test suites had to be slightly modified to
ensure consistent behavior. The previous artifical graph had only
one hypernode remaining that was not contained in any hyperedge.
Thus the new hypergraph implementation disabled that node automatically.
The addtion of a new hypernode and a new HE now ensures that node 0
remains enabled after the 3 single node HEs have been removed.