change hypergraph behavior: unconnected hypernodes are removed
Previous version generally left hypernodes enabled after removing the last
incident hyperedge. This behavior is intended during coarsening
(we wan't those hypernodes in the graph as they contain information
about how many other nodes had been contracted to this node -> weight).

With the upcoming introduction of cmaxnet-parameter, which determines
the size of hyperedges that will be sacrificed as cut edges in order to
boost up contraction, we however also want to really _remove_ a hyperedge
from the graph, i.e. including all nodes that are only connected to this
hyperedge. We therefore introduce a bool parameter to switch whether un-
connected hypernodes should be removed from the graph or not.

Coarsening/uncoarsening has this switch set to false. Cmaxnet-removal
will have set this switch to true.

If we do not remove unconnected hypernodes in the cmaxnet-case, this will
lead to the problem that certain hypergraphs (for example cyrstk01.hgr with
cmaxnet=50) will not be able to be coarsened up to the predifined minimum
number of hypernodes, because a large number of unconnected hypernodes
will remain in the graph after sacrificing hyperedges > 50 pins. All
of these then will return an invalid rating and remain in the coarsened
graph.

Testcases in hypergraph_test have been adapted to reflect the behavior
intended by the bool flag.