prepare rater for vcycles - acccept only if nodes are in same partition
We use the most straightforward way to implement vcycle-coarsening:
During the first coarsening run, each node is in the invalid partition.
Therefore the newly introduces condition alway evaluates to true.
In successive vcycle-coarsening iterations the graph is already partitioned.
In this case, we only accept a contraction pair, if both nodes are in the same
partition.