use bitvector to remember which HEs have been activated
previous version used PQ containment check. However we do not
want to do a containment check in both PQs each time we want to check
whether or not a HE has already been activated. Aditionally we would
have to check _both_ PQs, as it might be the case that a HE is only
present in one of them, because it was removed from the other one due to
clogging.