switch to public inheritance and CRTP to allow reuse of HE removal methods
Hyperedge-based coarseners need to remove single-node and parallel
Hyperedges from their priority queues after they are removed from the
hypergraph. Currently we reuse the code to detect and remove those
hyperedges from CoarsenerBase.

To avoid having to reiterate over all removed hyperedges after the
removal, just to remove those IDs from the PQ, we provide a CRTP-based
approach and allow subclasses to implement a method that deals with
removed HEs _while_ they are being removed. Hypernode-based coarseners
just have to define an empty method.

Negative side effect: We had to weaken inheritance from private to
public in order to be able to access the derived class's method.
Technically, private inheritance would be the right decision, as this
isn't really an 'is-a' relationship, but an 'is-implemented-in-terms-of'
relationship. However as of now, I'm willing to sacrifice this for
efficiency. In any case, partitioning clients should always use the provided
interface and the factory class.

This scheme will also be necessary when parallel HEs are removed.