Uncomment assertion for rating consistency
This assertion was introduced with commit
1da08861ac4fd2d920e01444884b29e98d6a4ff2
as part of a consolidation effort. However it is in the nature of the
heuristic heavy-edge coarsener to __not__ update the ratings of all
hypernodes. Instead it updates only the ratings for hypernodes that are
directly affected by the contraction (i.e. the hypernodes that had
either the representative or the contracted node as contraction target).
Therefore ratings might become slightly inconsistent during coarsening.