use connectivity sets for kway gain calculation - this might hurt balance
Problem: The previous version checked gains for movements into all
partitions (even into those that are not connected by an incident HE).
Thus this version was able to perform a decent amount of rebalancing in
case of many equal gain values.

The current version now only considers parts that are connected via an
indicent HE. This reduces the possibilities to perform rebalancing
operations.