reintroduces imbalanced partition constraint
Previous version allowed max_gain moves as soon as it
would reduce the difference in part sizes. this however leads
to incredibly long runtime for refiner. Thus we go back to the
initial version and allow same gain moves only if the parts are
above L_max and rebalancing is possible.