consolidate code with minor code change
Previous version only performed a move, if it was feasible and only
then updated the neighbours etc. This behavior is equivalent to first
finding the max gain move that is acutally feasible (and to discard all
unfeasible moves) and then performing that move.

Slight change that might have improved performance: Removed infeasible
moves were marked as moved before this commit, although they were never
moved. Now, only moves that are actually performed are marked as such.

The behavior of the new version is exactly the same as the old version,
if one only marks a move as _marked[hn]=true, if the move is actually
performed.