fix hyperedge coarsener for hypergraphs with edges of size 1
Hypergraph instances might contain hyperedges that only contain one pin.
In this case, no contraction is performed and therefore there is no
coarsening memento in this case. The corresponding
hypepredgecoarsener-memento contains only information about the removed
single-node hyperedge. However the previous version did not check for
the size of coarsening memento and therefore included invalid hypernodes
in the refinement process!