fix performance bug: redundant reRatings heuristic coarsener
Heuristic heavy edge coarsener showed poor performance on sparse matrix
instances. Profiling revealed that this was due to a high load on
reRateHypernodesAffectedByParallelHyperedgeRemoval. In case of SPM
instances, this was especially critical, because these instances produce
a lot of parallel hyperedges.

Starting with this commit we now evaluate every affected pin only once!