consolidate use of stopping rule and modify simple stopping rule
Previous version considered zero-gain moves as an improvement in some
refiners. We now only consider actual decrease in cut an improvement
consistently across all refiners. This might be changed in the future.

Furthermore, the old implementation of SimpleStoppingRule did perform
more than the specified amout of fruitless moves due to the way it was
implemented and used. The new implementation now really aborts after x
fruitless moves that did not lead do a decrease in cut.