fix KFM
2 major bug fixes:
- Previous version missed to insert possible moves if the connectivity
of a HE did increase after a move. This version now also inserts a move
to the newly connected part for all pins.
- In addition, the previous version missed to activate pins of a
hyperedge if it first became an internal HE (on state change free ->
loose we assumed to activate those pins). If another move later on made
the HE a cut HE again, then we did not insert the Pins into the PQ.