bugfix: global connectivity decrease
calculation of the global connectivity decrease for a potential move did
not account for the fact that if the HN to be moved is the only HN in
the source part for a HE, the assumed connectivity increase has to be
corrected by -1 for all other parts since a part-change won't increase
the connectivity (but this is assuemd by initializing each
decrease-counter with nodeDegree(hn)).

Additionally, this commit provides a new lambda-assertion to verify the
calculated connectivity decrease values by simulating the move, since
the original assertion contained the same bug as the calculation!