enhance KFM pq to handle enabling/disabling of heaps
The current version automatically enables each encountered part
and thus works exactly as the previous version with no significant
performance penalty.

The upcoming versions of Refiners will explicitly enable and disable
underloaded/overloaded parts. In this case, the PQ has to be adapted to
not enable heaps by default and the 2 disabled testcases can be activated.