add testcase to explicitely check pinCounts of restored HEs
Previous testcases did this implicitly. However, in the
multilevel context, we remove hyperedges that become
parallel or single-node. During uncoarsening theses HEs are
restored. Since they are "added" to the now partitioned hypergraph,
we have to maintain the connectivity, connectivitySet and pinCount
datastructures. While there are explicit tests for the first two,
the last one was only impicitly tested by another testcase.