introduce O(1) border-node check
Hypergraph now provides O(1) border node check
instead of O(deg(v)). This is achieved by maintaining
the number of incident cut HEs per Hypernode during
uncoarsening. This version does this inside the hypergraph
and works for 2FM as well as KFM variants.

In /patches, there also exists a version, in which the
cut weight of incident HEs is maintained in 2FM. However,
this version is slower because of additional maintenance overhead
due to the fact that we need to be able to rollback chached changes.