introduce EnhancedBucketQueue and refactor PQ-related tests
EnhancedBucketQueue is a more compley BucketQueue implementation
that allows O(1) clear combined with lazy-reset of buckets.
Additionally it uses a search tree to keep track of the non-empty
buckets. Its performance comes close to that of the NoDataBinaryHeap.

PriorityQueue test cases are now refactored and used typed tests
to test both implementations with one test case implementation.
This makes the playground/pq_test superflouus. However the current
tests do not test floating point key types as this is currently
not supported by EnhancedBucketQueue.