Tune fullUpdate for 2FM - lazy non-border node removal
On large hypergraphs like kkt_power, significant running time
went in checking whether or not a hypernode remained border
hypernode after an FM-move.
Now, the hypergraph DS is able to track all hypernodes that become
internal during changeNodePart and hand them back to the caller.
2FM then uses this to perform lazy border node removal AFTER all
updates are performed.
This allows us to get rid of of the border node check in fullUpdate
at the cost of updating nodes the will be deleted anyway (which actually
does not lead to any significant performance penalty).