allow gain budgeting for rebalancing
previous version only allowed rebalancing moves
if the rebalancing gain was positive. This is very
restrictive.
This is now a more relaxed variant that allows rebalancing
as long as the overall gain  (best_cut - current_cut) stays
at least 1. This effecively means that if we find a positive
gain move that makes the solution infeasible, we can still try
to restore feasibility bei sacrificing some of the positive gain.

A previous version also tried to rebalance aggressively (regardless
of the gain), i.e. it also allowed negative gain moves. However
this was too agressive and did not yield any improvements - most
of the tested results actually became worse.