BUGFIX: peek() accessed memory initialized to 0
Since peek() looks 'one element past the current size', it looks outside
the dense datastructure, if it is at its original size. Since we did not
sanitize the memory of dense, and since it is initialized to 0, a lazy
coarsener test failed.

More precicely, the uncontraction operation tried to determine whether
or not do perform a case 1 undo opertion for v=0 and peeked 'one
element past' the actual size. Since this was unused memory, it was
initialized to 0. Therefore the uncontraction operation decided to do
a case 1 undo instead of a case 2 undo.