fix bug in MSVC builds
Unfortunatly, MSVC did not do copy elision for Context objects.
This lead to the problem of context objects having dead parent
pointers, which in turn, caused the segfaults.
We mitigate this problem by directly referring to the main context
on copy construction. This can be done, since data will end up there
anyway.