allow to change k
In evolutionary HGP, we use cross combine operations that
combine solutions for different values of k. In order to reuse
the existing hypergraph, we need a way to change the value of k,
which the hypergraph initially was constructed for.
Method changeK(k) now allows exactly this and resizes internal
data structures accordingly.