refactor stats
Stats are now disabled by default and can be enabled via GATHER_STATS macro
in definitions.h. Stats handling code will be optimized away in case
GATHER_STATS is not defined.

Stats class is now also more generic. Adding a new Stats checkpoint is
as simple as adding a value to the enum class and a corresponding case
to operator <<.