Bugfix: unused communities in case of sparsification
This fixes a bug introduced in commit
901176a6fe78ccd756af84c9adba9867d63e1b26 .

Since community detecion was moved into the preprocessing method of
the partitioner, it was performed directly after single node net
removal. Unfortunately, this meant that in case of sparsifiacation,
community detection was performed on the original hypergraph and then
sparsification build the new sparsified hypergraph (with empty
community structure), which was then partitioned.

Along with the bugfix, this commit also contains assertions and a
regression testcase, which ensure that community information is used
corrently if the sparsifier is enabled.