fix violation of LazyCoarsener invariant due to cmaxnet
By adding the cmaxnet parameter for ml-style coarsening, we introduced
the a bug in the lazy coarsener, which violated it's main invariant
that for any two possible contraction partners both are in the priority queue.
Using cmaxnet can lead to a coarsening process that violates
this invariant in the following way:
Assume that v0 is a vertex incident two two nets. Let n0 be a net smaller
than cmaxnet and n1 a net just above the cmaxnet threshold. Further
assume that n1 contains only one pin v1 (also incident to n0)
light enough to be contracted with v0. All other pins of n1 are to heavy,
while all other pins of n0 are light enough to be contracted
with v0/v1.
Because of this setup, the only possible contraction is (v0,v1), since
all other pins of n1 are to heavy and the pins of n0 are ignored during
rating because |n0| > cmaxnet.
Now, the LazyCoarsener starts by contracting (v0,v1).
After contracting (v0,v1), |n0| <= cmaxnet.
This leads to new possible contractions of v0 with pins of n0.
However, since these pins have been ignored previously, none of the is in
the priority queue. Therefore assertion
ASSERT(_pq.contains(contracted_node), V(contracted_node)); fails!

This commit fixes this bug and includes a testcase.