Bugfix: Sparsifier ignored vertex / hyperedge weights
If the original hypergraph contained either vertex or hyperedge
weights, the sparsifier did not consider them when building the
sparsified hypergraph. This lead to grossly imbalanced partitions,
since the partitioner itself the works with and 'almost' unweighted
graph. 'Almost unweighted', since the sparsifier did count contracted
vertices and parallel edges. However only these counts were used as
weights.

This commit fixes the bug and adds a test case with an instance that
contains both vertex and edge weights.

@yarchi:The current sparsifier configuration is problematic for
vertex-weighted hypergraphs, because of the
'--p-sparsifier-max-cluster-size' parameter. This is currently set to
10 (allowing only small clusters). For hypergraphs with vertex weights
> 10, this leads to the fact that the sparsifier doesn't do anything
except wasting running time. How can we fix this?