improve community detection for graphs
Previously, KaHyPar treated hypergraphs that were actually
graphs strictly as hypergraphs during community detection, i.e., it
created a bipartite graph of the graph -- blowing up both the number
of nodes and the number of edges and changing the graph structure.

With this commit, KaHyPar now recognizes graphs and uses the louvain
algorithm directly on the graph structure.

This speeds up community detection for graphs significantly and might
also be benefical for the overall partitioning process in terms of
solution quality, since the Louvain algorith does not unnecessarily
have to work on the bipartite graph.