fix interaction of sparsifier with deduplicator
The sparsifier implicitly assumed to be the first algorithm
to be called on the hypergraph and therefore did not use the correct
node iterators as well as size functions to get the current number of
vertices and nets. This lead to incorrect results when both
the sparsifier and the deduplicator were active.