fix gain range for bucket priority queues
KaHyPar supports both bucket- and binary-heap-based priority queues.
Heap-based PQs are the default. This commit fixes a bug for
bucket-based PQs: The maximum vertex gain was estimated to be the
maximum degree times the maximum hyperege weight present in the
coarsest hypergraph. However, removed single-node hyperedges can
actually have a larger weight, in which case the gain range is not
calculated correctly.
We now take the maximum of the heaviest removed single-node hyperedge
and the heaviest hyperedge present in the coarsest hypergraph as new estimate.