reduce memory consumption of connectivity sets
Thanks to the gain cache, the connectivity sets data structure
is rarly used within KaHyPar. The last time we reduced its size,
we gave up contant time containment checks and removals. Therefore
it is not necessary anymore to actually allocate the structure for
all nets.

This commit now simplifies the data structure. Instead of one _large_
array, we now just use a vector of vectors.