Modified hypergraph (#42)
* First step towards more space efficient hypergraph

The initial hypergraph data strucuture used a single vector to store
both the incident nets of all vertices and the pins of all edges.
However, during coarsening, the former can grow, while the latter only
shrinks. In the previous version, increasing the degree of a vertex
lead to copies of all of its incidenet nets. This was previously
necessary in order to be able to restore incidence information during
uncontractios.

Here, we introduce a new way to handle the problem, by storing
incident nets in a single vector per vertex and use sentinel
information during uncontractions.

The proposed data structure modification is currently under testing
and will be merged into master once the performance tests are done.

* cleanup