fix FM vertex activation for individual block weights
Both k-way FM implementation used the weight of block 0 to decide
whether or not to activate a vertex. While this is OK in the default
setting (where the weights of all blocks are equal), we have
to use specific block weights when KaHyPar is configured
to use individual block weights.

This commit fixes the vertex actication along with the corresponding
assertions.