Make modularity computation robust to graphs without edges
When a bipartite graph doesn't have any edges, the previous
code resulted in a division by 0. Now, the computation works
as expected and every vertex is assigned its own community.

Thanks to Nikolai Maas for implicitly catching this through
the introduction of heavy assertions and thanks to Robin
Andre for debugging this issue.