Fixed Vertices with Individual Part Weights (Fix Issue #119) (#121)
* Fix Issue #119

* Fix Issue #119:
We use a maximum weighted bipartite matching algorithm to compute an assignment of block ids to the fixed vertex blocks that results in the least increase in the objective function. If an assigment of a block to fixed vertex block results in an imbalanced partition, we set the weight of the corresponding edge in the bipartite graph to zero. However, the maximum weighted bipartite matching algorithm assumes that it operates on a complete graph and therefore matchings containing edges with zero weight are still feasible. The fix sets weight of these edges (which would result in an imbalanced partition) to an appropriately high negative weight such that a maximum weighted bipartite matching will not contain these edges.